+++ server/src/controllers/bookingController.js
@@ -6,73 +6,111 @@ class bookingController {
   static createBooking(req, res) {
     const decodedUser = req.user;
     if (decodedUser) {
-      const { trip_id } = req.body;
+      const {
+        trip_id
+      } = req.body;
+
       const booking = {
         trip_id,
         created_on: new Date()
       };
 
-      const tripQuery = {
-        text: 'SELECT bus_id, trip_date FROM trips WHERE id = $1',
-        values: [booking.trip_id]
-      };
-
-      pool.query(tripQuery, (error, data) => {
-        if (data) {
-          const seatNumberQuery = {
-            text: 'SELECT * FROM bookings where trip_id = $1',
+      // get user details
+      const getUser = {
+        text: 'SELECT first_name, last_name, email FROM users WHERE id = $1',
+        values: [decodedUser.id]
+      }
+      pool.query(getUser, (err, userData) => {
+        if (userData) {
+          // Get bus_id & trip_date from trips table using trip_id
+          const tripQuery = {
+            text: 'SELECT bus_id, trip_date FROM trips WHERE id = $1',
             values: [booking.trip_id]
           };
-          pool.query(seatNumberQuery, (seatNumberError, seatNumberData) => {
-            let seatNumber = seatNumberData.rowCount;
-            seatNumber += 1;
+          pool.query(tripQuery, (error, data) => {
+            // Check Bus Sitting Capacity
+            const busSeatCapacity = {
+              text: 'SELECT capacity from buses WHERE id = $1',
+              values: [data.rows[0].bus_id]
+            }
 
-            const bookingQuery = {
-              text:
-                'INSERT INTO bookings (user_id, trip_id, bus_id, trip_date, seat_number, first_name, last_name, email, created_on) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) returning *',
-              values: [
-                decodedUser.id,
-                booking.trip_id,
-                data.rows[0].bus_id,
-                data.rows[0].trip_date,
-                seatNumber,
-                decodedUser.first_name,
-                decodedUser.last_name,
-                decodedUser.email,
-                booking.created_on
-              ]
-            };
-            pool.query(bookingQuery, (bookingError, bookingData) => {
-              return res.status(201).send({
-                status: 'success',
-                data: {
-                  id: bookingData.rows[0].id,
-                  user_id: bookingData.rows[0].user_id,
-                  trip_id: bookingData.rows[0].trip_id,
-                  bus_id: bookingData.rows[0].bus_id,
-                  trip_date: bookingData.rows[0].trip_date,
-                  seat_number: bookingData.rows[0].seat_number,
-                  first_name: bookingData.rows[0].first_name,
-                  last_name: bookingData.rows[0].last_name,
-                  email: bookingData.rows[0].email
+            pool.query(busSeatCapacity, (err, busCapacity) => {
+
+              // Check for current bookings per trip
+              const currentBookings = {
+                text: 'SELECT * FROM bookings WHERE trip_id = $1',
+                values: [booking.trip_id]
+              }
+              pool.query(currentBookings, (currentBookingError, currentBookingData) => {
+                if (currentBookingData.rowCount < busCapacity.rows[0].capacity) {
+                  // Get all bookings where trip_id = $1
+                  const seatNumberQuery = {
+                    text: 'SELECT * FROM bookings where trip_id = $1',
+                    values: [booking.trip_id]
+                  };
+                  pool.query(seatNumberQuery, (seatNumberError, seatNumberData) => {
+                    let seatNumber;
+                    if (seatNumberData.rowCount !== 0) {
+                      seatNumber = seatNumberData.rowCount;
+                      seatNumber += 1;
+                    } else {
+                      seatNumber = 1;
+                    }
+                    const bookingQuery = {
+                      text: 'INSERT INTO bookings (user_id, trip_id, bus_id, trip_date, seat_number, created_on) VALUES ($1, $2, $3, $4, $5, $6) returning *',
+                      values: [
+                        decodedUser.id,
+                        booking.trip_id,
+                        data.rows[0].bus_id,
+                        data.rows[0].trip_date,
+                        seatNumber,
+                        booking.created_on
+                      ]
+                    };
+                    pool.query(bookingQuery, (bookingError, bookingData) => {
+                      return res.status(201).json({
+                        status: 'success',
+                        data: {
+                          id: bookingData.rows[0].id,
+                          user_id: bookingData.rows[0].user_id,
+                          trip_id: bookingData.rows[0].trip_id,
+                          bus_id: bookingData.rows[0].bus_id,
+                          trip_date: bookingData.rows[0].trip_date,
+                          seat_number: bookingData.rows[0].seat_number,
+                          first_name: userData.rows[0].first_name,
+                          last_name: userData.rows[0].last_name,
+                          email: userData.rows[0].email
+                        }
+                      });
+                    });
+                  });
+                } else {
+                  res.status(400).json({
+                    status: 'error',
+                    error: 'Sorry Bus is Full - Bus Capacity cannot be Exceeded'
+                  })
                 }
-              });
-            });
+              })
+            })
           });
         }
-      });
+      })
+    } else {
+      return res.status(400).json({
+        status: 'error',
+        error: 'Create Booking Failed!'
+      })
     }
   }
 
   static getBooking(req, res) {
     const decodedUser = req.user;
-
-    switch (decodedUser.is_admin) {
+    switch (decodedUser.isAdmin) {
       case true:
         const adminQuery = 'SELECT * from bookings';
 
         pool.query(adminQuery, (error, data) => {
-          return res.status(200).send({
+          return res.status(200).json({
             status: 'success',
             data: [data.rows]
           });
@@ -81,18 +119,18 @@ class bookingController {
       case false:
         const nonAdminQuery = {
           text: 'SELECT * FROM bookings WHERE user_id = $1',
-          values: [decodedUser.user_id]
+          values: [decodedUser.id]
         };
 
         pool.query(nonAdminQuery, (error, data) => {
-          return res.status(200).send({
+          return res.status(200).json({
             status: 'success',
             data: [data.rows]
           });
         });
         break;
       default:
-        return res.status(200).send({
+        return res.status(200).json({
           status: 'error',
           error: 'Not Allowed'
         });
@@ -104,36 +142,55 @@ class bookingController {
 
     if (decodedUser) {
       if (Number.isNaN(req.params.bookingId)) {
-        return res.status(400).send({
+        return res.status(400).json({
           status: 'error',
           error: 'Invalid Booking ID'
         });
       }
-      const deleteQuery = {
-        text: 'DELETE FROM bookings WHERE id = $1',
+
+      // Check if booking has already been deleted
+      const checkBooking = {
+        text: "SELECT id FROM bookings WHERE id = $1",
         values: [req.params.bookingId]
       };
 
-      pool.query(deleteQuery, (error, data) => {
-        if (data.rows.length === 0) {
-          return res.status(200).send({
-            status: 'success',
-            data: {
-              message: 'Booking deleted successfully'
-            }
-          });
+      pool.query(checkBooking, (error, data) => {
+        if (data.rowCount === 0) {
+          return res.status(409).json({
+            status: 'error',
+            error: 'Booking has already been deleted'
+          })
         }
-      });
+
+        const deleteQuery = {
+          text: 'DELETE FROM bookings WHERE id = $1',
+          values: [req.params.bookingId]
+        };
+
+        pool.query(deleteQuery, (err, data) => {
+          if (data.rows.length === 0) {
+            return res.status(200).json({
+              status: 'success',
+              data: {
+                message: 'Booking deleted successfully'
+              }
+            });
+          }
+        });
+      })
     }
+
   }
 
   static changeSeatNumber(req, res) {
     const decodedUser = req.user;
-    const { seat_number } = req.body;
+    const {
+      seat_number
+    } = req.body;
 
     if (decodedUser) {
       if (Number.isNaN(req.params.bookingId)) {
-        return res.status(400).send({
+        return res.status(400).json({
           status: 'error',
           error: 'Invalid Booking ID'
         });
@@ -152,32 +209,46 @@ class bookingController {
           };
 
           pool.query(checkSeatQuery, (seatError, seatData) => {
-            const filteredData = seatData.rows.filter(
-              seatNumber => seatNumber.seat_number === seat_number
-            );
-
-            if (filteredData.length > 0) {
-              return res.status(409).send({
-                status: 'error',
-                error: 'Seat Number has been taken'
-              });
+            // Check Bus Capacity
+            const busCapacity = {
+              text: 'SELECT capacity from buses WHERE id = $1',
+              values: [seatData.rows[0].bus_id]
             }
+            pool.query(busCapacity, (seatCapacityError, seatCapacityData) => {
+              if (seat_number > seatCapacityData.rows[0].capacity) {
+                res.status(400).json({
+                  status: 'error',
+                  error: 'Number exceeds the Bus sitting capacity'
+                })
+              } else {
+                const filteredData = seatData.rows.filter(
+                  seatNumber => seatNumber.seat_number === seat_number
+                );
 
-            const changeSeatNumberQuery = {
-              text: `UPDATE bookings SET seat_number = ${seat_number} WHERE id = $1`,
-              values: [req.params.bookingId]
-            };
+                if (filteredData.length > 0) {
+                  return res.status(409).json({
+                    status: 'error',
+                    error: 'Seat Number has been taken'
+                  });
+                }
 
-            pool.query(changeSeatNumberQuery, (newError, newData) => {
-              if (newData) {
-                return res.status(200).send({
-                  status: 'success',
-                  data: {
-                    message: 'Seat Number changed successfully'
+                const changeSeatNumberQuery = {
+                  text: `UPDATE bookings SET seat_number = ${seat_number} WHERE id = $1`,
+                  values: [req.params.bookingId]
+                };
+
+                pool.query(changeSeatNumberQuery, (newError, newData) => {
+                  if (newData) {
+                    return res.status(200).json({
+                      status: 'success',
+                      data: {
+                        message: 'Seat Number changed successfully'
+                      }
+                    });
                   }
                 });
               }
-            });
+            })
           });
         }
       });
